#1396 Moved extraction of Letter object to a dedicated class
